Update to version 0.95. * added specific XNAT tag context menu * set tag value dialog auto focus * auto-disable "upload" menu on FSfolders. * Support for complex tag clauses (REST path parts) in the Uploader. * Updated the modify tag value rule to support multiple capt. groups * Modeless progress dialog in the Uploader * Detailed comments in xml ontologies. *

+++ src/org/nrg/xnd/rules/RuleManager.java
@@ -30,11 +30,12 @@ public abstract class RuleManager
 			loadRule(f, bMacro, false);
 
 		// pre-existing custom rules.
+/*		
 		rf = new File(Utils.GetPluginPath() + "/xml_resources/customRules");
 		files = rf.listFiles();
 		for (File f : files)
 			loadRule(f, bMacro, false);
-
+*/
 		// pre-existing default rules.
 		for (int i = 0; i < Rule.DEFAULT_RULES.length; i++)
 		{
@@ -75,8 +76,7 @@ public abstract class RuleManager
 	{
 		Rule r = m_rules.get(name);
 		File f = r.getFile();
-		if (f != null)
-			f.delete();
+		if (f != null) f.delete();
 		m_rules.remove(name);
 	}
 	public static TreeMap<String, Rule> getRules()
